database/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.13.2
python-dotenv==1.1.1
cloudinary==1.41.0
python-multipart==0.0.20
